refactor: disable game scheduling, split dashboard service
- scheduling service still exists but is disabled until a use is
  determined for it
- use of the game schedule on the client was removed

- dashboard service/routes and client-side dashboard components removed or
  split into logical components for current or future use. specifically,
  player stats were moved to a leaderboard service and (currently
  unused) code dealing with surveys was given a survey service
