+++ client/src/components/global/Schedule.vue
