+++ client/src/components/global/Sponsors.vue
