+++ server/deploy/settings.template.json
@@ -1,13 +1,11 @@
 {
   "isTournamentSignUpEnabled": false,
   "isFreePlayEnabled": true,
-  "isAutoSchedulerEnabled": true,
+  "isAutoSchedulerEnabled": false,
   "autoSchedulerHourInterval": 3,
   "autoSchedulerDaysOut": 1,
   "maxConnections": 200,
   "defaultDaysMuted": 3,
-  "lobbyGroupAssignmentInterval": 1,
-  "lobbyForceGroupAssignmentInterval": -1,
   "lobbyOpenBeforeOffset": 10,
-  "lobbyOpenAfterOffset": 5
+  "lobbyOpenAfterOffset": 1
 }
