--- server/src/routes/dashboard.ts
@@ -1,27 +1,0 @@
-import { Router, Request, Response } from 'express';
-import { getServices } from '@port-of-mars/server/services';
-import { User } from '@port-of-mars/server/entity/User';
-import { isAuthenticated } from '@port-of-mars/server/routes/middleware';
-export const dashboardRouter = Router();
-
-dashboardRouter
-  .get('/', isAuthenticated, async (req: Request, res: Response, next) => {
-    try {
-      const dashboardData = await getServices().dashboard.getData(req.user as User);
-      res.json(dashboardData);
-    }
-    catch (e) {
-      next(e);
-    }
-  });
-
-dashboardRouter
-  .get('/lobby-status', isAuthenticated, async (req: Request, res: Response, next) => {
-    try {
-      const isLobbyOpen = await getServices().dashboard.isLobbyOpen(req.user as User);
-      res.json(isLobbyOpen);
-    }
-    catch (e) {
-      next(e);
-    }
-  });
